fix: resolve ESLint warnings and ts errors (useEffect dependency warnings fixed)
