+++ src/App.tsx
@@ -4,7 +4,7 @@ import RouterPage from '@/router/index';
 import { RecoilRoot } from 'recoil';
 import RootLayout from '@/pages/RootLayout';
 import { GoogleOAuthProvider } from '@react-oauth/google';
-import AuthProvider from '@/context/AuthContext';
+import AuthProvider from '@/context/authProvider';
 
 export default function App() {
     return (
